chang the location of the workload benchmark test
